типизация в отдельном файле (#8)
